Update Ukrainian in BlueskyClient.Uwp.uk.xlf (№2)
I rewrote some words and sentences in the xlf file))), and where I could, I checked for inaccuracies. In addition to not being able to do everything, I also need to do something with the following: where my profile is, there is a post, followers, and subscriptions. In Ukrainian, there is a singular and plural, I replaced it with the plural, but it will be incorrect when there are 1 дописів, if there should be 1 допис, etc.
For 1: 1 допис, 1 підписник, 1 підписка.
For 2: 2 дописа, 2 підписника, 2 підписки
For 3 and up, what I have already edited.